only hitting canvas when things change (all draw primitives + frame)

+++ pax-message/src/lib.rs
@@ -554,7 +554,7 @@ pub struct CheckboxStyleMessage {
 #[repr(C)]
 pub struct OcclusionPatch {
     pub id: u32,
-    pub occlusion_layer_id: u32,
+    pub occlusion_layer_id: usize,
     pub z_index: i32,
     pub parent_frame: Option<u32>,
 }
